Handle DVDNAV title changes
DVDNAV reports title changes as VTS changes, which leaves us with the job
to map from VTS to title. When the number of titles and VTS's match, the
VTS number is used, otherwise the duration of the title and hints from the
gui are used to try and detect the mapping from VTS to title.